+++ src/components/Messaging.jsx
@@ -2,13 +2,32 @@ import React, {Component} from 'react'
 import db from '../firestore'
 import Message from './Message'
 import firebase from 'firebase'
+import {DragSource} from 'react-dnd'
 
 import PropTypes from 'prop-types'
 import {withStyles} from '@material-ui/core/styles'
-import classNames from 'classnames'
 import Button from '@material-ui/core/Button'
+import Card from '@material-ui/core/Card'
+import CardContent from '@material-ui/core/CardContent'
+import Typography from '@material-ui/core/Typography'
+
+const messagingSource = {
+  beginDrag(props) {
+    return {}
+  }
+}
+
+function collect(connect, monitor) {
+  return {
+    connectDragSource: connect.dragSource(),
+    isDragging: monitor.isDragging()
+  }
+}
 
 const styles = theme => ({
+  card: {
+    minWidth: 275
+  },
   button: {
     margin: theme.spacing.unit
   },
@@ -79,39 +98,62 @@ export class Messaging extends Component {
   }
 
   render() {
-    const {classes} = this.props
-    return (
-      <div id="messages">
-        <div>
-          {this.state.messages.map(message => (
-            <Message key={message.id} message={message} />
-          ))}
-        </div>
-        <form onSubmit={this.handleSubmit} className="columns modal-card-foot">
-          <input
-            type="text"
-            name="newMessage"
-            value={this.state.newMessage}
-            onChange={this.handleChange}
-          />
+    const {classes, connectDragSource, isDragging} = this.props
+    return connectDragSource(
+      <div>
+        <Card
+          className={classes.card}
+          style={{
+            opacity: isDragging ? 0.5 : 1,
+            cursor: 'move'
+          }}
+        >
+          <CardContent>
+            <Typography className={classes.title} color="textSecondary">
+              Chat
+            </Typography>
+            <div id="messages">
+              <div>
+                {this.state.messages.map(message => (
+                  <Message key={message.id} message={message} />
+                ))}
+              </div>
+              <form
+                onSubmit={this.handleSubmit}
+                className="columns modal-card-foot"
+              >
+                <input
+                  type="text"
+                  name="newMessage"
+                  value={this.state.newMessage}
+                  onChange={this.handleChange}
+                />
 
-          <Button
-            type="submit"
-            size="small"
-            variant="outlined"
-            color="primary"
-            className={classes.button}
-          >
-            Send
-          </Button>
-        </form>
+                <Button
+                  type="submit"
+                  size="small"
+                  variant="outlined"
+                  color="primary"
+                  className={classes.button}
+                >
+                  Send
+                </Button>
+              </form>
+            </div>
+          </CardContent>
+          <Button>Remove</Button>
+        </Card>
       </div>
     )
   }
 }
 
 Messaging.propTypes = {
-  classes: PropTypes.object.isRequired
+  classes: PropTypes.object.isRequired,
+  connectDragSource: PropTypes.func.isRequired,
+  isDragging: PropTypes.bool.isRequired
 }
 
-export default withStyles(styles)(Messaging)
+export default DragSource('MODULE', messagingSource, collect)(
+  withStyles(styles)(Messaging)
+)
